pyproject.toml
[project]
name = "usst-rollcall"
version = "0.1.0"
description = "USST TronClass rollcall watcher and notifier"
readme = "README.md"
authors = [
    { name = "Nitmi", email = "0@123778.xyz" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
    "typer>=0.24.2",
]

[project.scripts]
usst-rollcall = "usst_rollcall.cli:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"
